refactor: restructure into a Deno workspace monorepo
Move the @zuke/core package to packages/core ahead of the tool wrapper
packages (@zuke/deno, @zuke/npm, @zuke/cmd) that will join the workspace.
