+++ deno.json
@@ -1,15 +1,13 @@
 {
-  "name": "@zuke/core",
-  "version": "0.0.0",
-  "exports": {
-    ".": "./mod.ts",
-    "./shell": "./src/shell.ts"
+  "workspace": ["packages/core"],
+  "imports": {
+    "@zuke/core": "jsr:@zuke/core@^0.0.0"
   },
   "tasks": {
     "test": "deno test -A",
     "cov": "deno test -A --coverage=cov_profile && deno coverage cov_profile --lcov '--exclude=(tests|scripts)/' --output=cov.lcov && deno run --allow-read scripts/check-coverage.ts cov.lcov 95",
     "cov:report": "deno coverage cov_profile '--exclude=(tests|scripts)/'",
-    "check": "deno check mod.ts src/*.ts tests/*.ts scripts/*.ts zuke.ts",
+    "check": "deno check zuke.ts scripts/*.ts tests/*.ts packages/*/mod.ts packages/*/src/*.ts packages/*/tests/*.ts",
     "fmt": "deno fmt",
     "fmt:check": "deno fmt --check",
     "lint": "deno lint",
@@ -17,10 +15,10 @@
     "ci": "deno fmt --check && deno lint && deno task spell && deno task check && deno task cov"
   },
   "fmt": {
-    "include": ["mod.ts", "zuke.ts", "src/", "tests/", "scripts/"]
+    "include": ["zuke.ts", "scripts/", "tests/", "packages/"]
   },
   "lint": {
-    "include": ["mod.ts", "zuke.ts", "src/", "tests/", "scripts/"],
+    "include": ["zuke.ts", "scripts/", "tests/", "packages/"],
     "rules": {
       "tags": ["recommended"],
       "include": ["no-explicit-any", "verbatim-module-syntax", "ban-untagged-todo"]
