+++ deno.lock
@@ -554,5 +554,10 @@
       "integrity": "sha512-2AvhNX3mb8zd6Zy7INTtSpl1F15HW6Wnqj0srWlkKLcpYl/gMIMJiyuGq2KeI2YFxUPjdlB+3Lc10seMLtL4cA==",
       "bin": true
     }
+  },
+  "workspace": {
+    "dependencies": [
+      "jsr:@zuke/core@^0.0.0"
+    ]
   }
 }
