+++ packages/core/src/cli.ts
