+++ packages/core/src/shell.ts
