+++ packages/core/tests/cli_test.ts
