+++ packages/core/tests/shell_test.ts
